+++ pkgs/native_toolchain_c/test/native_toolchain/xcode_test.dart
@@ -8,6 +8,7 @@ library;
 
 import 'dart:io';
 
+import 'package:file/local.dart';
 import 'package:native_toolchain_c/src/native_toolchain/xcode.dart';
 import 'package:native_toolchain_c/src/tool/tool.dart';
 import 'package:native_toolchain_c/src/tool/tool_instance.dart';
@@ -61,6 +62,7 @@ void main() {
       tool: tool,
       logger: logger,
       processManager: const LocalProcessManager(),
+      fileSystem: const LocalFileSystem(),
     );
     expect(result, <ToolInstance>[]);
   });